+++ CHANGELOG.md
@@ -1,5 +1,7 @@
+# CHANGELOG
+
 # Release 3.1.0-SNAPSHOT 
- * TODO
+ * Add: structural equality check for `NormalizedJsonPath` and `NormalizedJsonPathSegment` 
 
 # Release 3.0.1
  * Fix: Validates whether the whole string has been consumed